+++ cardinal_pythonlib/psychiatry/drugs.py
@@ -4,7 +4,7 @@
 """
 ===============================================================================
 
-    Original code copyright (C) 2009-2018 Rudolf Cardinal (rudolf@pobox.com).
+    Original code copyright (C) 2009-2019 Rudolf Cardinal (rudolf@pobox.com).
 
     This file is part of cardinal_pythonlib.