+++ docs/source/conf.py
@@ -25,7 +25,7 @@ from cardinal_pythonlib.version_string import VERSION_STRING
 
 project = 'cardinal_pythonlib'
 # noinspection PyShadowingBuiltins
-copyright = '2009-2018, Rudolf Cardinal'
+copyright = '2009-2019, Rudolf Cardinal'
 author = 'Rudolf Cardinal'
 
 # The short X.Y version